\name{nutshellDE-package}
\alias{nutshellDE-package}
\alias{nutshellDE-Paket}
\alias{nutshellDE}
\docType{package}
\title{
Beispieldaten zu  R in a Nutshell [deutsche Ausgabe]
}
\description{
Dieses Paket enthält alle Datensätze und eine kleine Auswahl 
Hilfsfunktionen, die als Beispiele in \emph{R in a Nutshell} 
[deutsche Ausgabe] des O'Reilly Verlags verwendet werden 
(ISBN 978-3-89721-649-5). Die komprimierte Paketdatei ist rund 
30 MB groß, das entpackte und installierte Paket belegt etwa 
60 MB Plattenspeicher. 
}
\details{
\subsection{Urheberrechte und Lizenzierung}{
Die \emph{Audioscrobbler}-Daten (Stand 6. Mai 2005) unterliegen der 
folgenden Creative-Commons-Lizenz: \cr
  \url{http://creativecommons.org/licenses/by-nc-sa/1.0/} 

Für alle anderen Datendateien gilt die Creative-Commons-Lizenz 
mit Namensnennung (CC BY 3.0, USA): \cr
  \url{http://creativecommons.org/licenses/by/3.0/us/}

Ein Teil der enthaltenen Funktionen wurde von Jörg Beyer erstellt. 
Diese werden (aus Kompatibilitätsgründen mit dem Rest des Pakets) 
unter Creative-Commons-Lizenz mit Namensnennung sowie Weitergabe 
unter gleichen Bedingungen gestellt (CC BY-SA 3.0, Deutschland): \cr
  \url{http://creativecommons.org/licenses/by-sa/3.0/de/} \cr
Die entsprechenden Inhalte sind in der Dokumentation kenntlich 
gemacht. 
}
\subsection{Änderungen}{
\ldots an den Datensätzen der amerikanischen Originalausgabe 
beschränken sich auf Eindeutschungen (Datensatz- und Variablennamen, 
Datenbestand), sowie auf Korrekturen technischer und inhaltlicher 
Fehler, die dort identifiziert wurden. Detailliertere Informationen 
sind in der Datei \code{ChangeLog} protokolliert. 
}
\subsection{Konventionen}{
Wir sind der Auffassung, dass die Nachvollziehbarkeit der Beispiele im Buch 
profitiert, wenn Daten -- soweit möglich und vernünftig -- ins Deutsche 
übertragen werden. 

Aus der Lokalisierung der Beispieldaten ergibt sich allerdings die Frage, 
welchen Aufwand es für den Anwender bedeutet, bei Bedarf eine Verbindung 
zwischen lokalisierten und Originaldaten herzustellen (um bspw. die 
Originaldokumentation nutzen zu können). Um diesen Schritt nicht unnötig 
zu verkomplizieren, haben wir in der Dokumentation der Datensätze die 
englischen Originalbezeichnungen in eckigen Klammern hinter die deutschen 
Spaltennamen gesetzt. 

Die Datenbasis, aus der wir die Beispiele aufbereitet haben, war 
in verschiedenen Fällen ausgesprochen umfangreich. 
Wenn wir im Folgenden auf die Bezugsquelle eines Datensatzes hinweisen, 
merken wir gegegebenfalls an, wenn der betreffende Download ungewöhnlich 
groß ist (falls Sie Interesse bekommen, sich mit den Quelldaten selbst 
zu befassen). 
}
}
\note{
Das R-Dokumentationsformat ist leider nicht lokaliserbar. 
Gliederungsüberschriften sind weitgehend festgelegt und in Englisch. 
}
\author{
Joseph Adler (Autor des nutshell-Pakets); 
Jörg Beyer (deutsche Übersetzung und Bearbeitung; ergänzender Code). 
}
\references{
Joseph Adler (Dez. 2010), \emph{R in a Nutshell} [deutsche Ausgabe]. 
O'Reilly Verlag, Köln. ISBN 978-3-89721-649-5, 
\url{http://www.oreilly.de/catalog/rinanutger/index.html}.

Der vollständige Beispielcode zum Buch ist unter der gleichen Adresse 
als Download verfügbar. 
}
\examples{
citation("nutshellDE")
}
\keyword{package}
